+++ LiteDB.Tests/Document/Bson_Tests.cs
@@ -29,7 +29,7 @@ namespace LiteDB.Tests.Document
             doc["Items"].AsArray.Add(new BsonDocument());
             doc["Items"].AsArray[0].AsDocument["Qtd"] = 3;
             doc["Items"].AsArray[0].AsDocument["Description"] = "Big beer package";
-            doc["Items"].AsArray[0].AsDocument["Unit"] = (double) 10 / (double) 3;
+            doc["Items"].AsArray[0].AsDocument["Unit"] = 10 / (double) 3;
 
             doc["Items"].AsArray.Add("string-one");
             doc["Items"].AsArray.Add(null);
